Organize doc site components into folders now that there are a lot of them
