+++ packages/docs/content/utilities/text/text-color.mdx
@@ -2,8 +2,6 @@
 title: Text Color
 ---
 
-import ColorSwatchList from '../../../src/components/ColorSwatchList';
-
 Use the text color utility to change the default text color of an element.
 
 The options below offer color palette combinations that fall within the range of Section 508 compliant and WCAG 2.0 level AA standards which is a minimum color contrast ratio of 4.5:1.
