+++ packages/docs/src/components/ContentRenderer/ColorSwatchList.tsx
