+++ packages/docs/src/components/ContentRenderer/ComponentThemeOptions.tsx
