+++ packages/docs/src/components/ContentRenderer/ContentRenderer.tsx
