+++ packages/docs/src/components/ContentRenderer/EmbeddedExample.tsx
