+++ packages/docs/src/components/ContentRenderer/ResponsiveExample.tsx
