+++ packages/docs/src/components/ContentRenderer/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ContentRenderer';
