+++ packages/docs/src/components/Icons/EmailIcon.tsx
