+++ packages/docs/src/components/Icons/GithubIcon.tsx
