+++ packages/docs/src/components/Icons/SlackIcon.tsx
