+++ packages/docs/src/components/Layout/ThemeSwitcher.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Dropdown } from '@cmsgov/design-system';
-import useTheme from '../helpers/useTheme';
-import { setQueryParam } from '../helpers/urlUtils';
+import useTheme from '../../helpers/useTheme';
+import { setQueryParam } from '../../helpers/urlUtils';
 
 const themeOptions = [
   { label: 'Core', value: 'core' },
